refactor(config): move syslogconf to config/syslog package (#1865)
