+++ config/config.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 package config
 
 import (
@@ -11,6 +9,7 @@ import (
 	"github.com/asaskevich/govalidator"
 	"golang.org/x/xerrors"
 
+	"github.com/future-architect/vuls/config/syslog"
 	"github.com/future-architect/vuls/constant"
 	"github.com/future-architect/vuls/logging"
 )
@@ -50,7 +49,7 @@ type Config struct {
 	Slack      SlackConf      `json:"-"`
 	EMail      SMTPConf       `json:"-"`
 	HTTP       HTTPConf       `json:"-"`
-	Syslog     SyslogConf     `json:"-"`
+	Syslog     syslog.Conf    `json:"-"`
 	AWS        AWSConf        `json:"-"`
 	Azure      AzureConf      `json:"-"`
 	ChatWork   ChatWorkConf   `json:"-"`
