+++ config/config_test.go
@@ -6,65 +6,6 @@ import (
 	. "github.com/future-architect/vuls/constant"
 )
 
-func TestSyslogConfValidate(t *testing.T) {
-	var tests = []struct {
-		conf              SyslogConf
-		expectedErrLength int
-	}{
-		{
-			conf:              SyslogConf{},
-			expectedErrLength: 0,
-		},
-		{
-			conf: SyslogConf{
-				Protocol: "tcp",
-				Port:     "5140",
-			},
-			expectedErrLength: 0,
-		},
-		{
-			conf: SyslogConf{
-				Protocol: "udp",
-				Port:     "12345",
-				Severity: "emerg",
-				Facility: "user",
-			},
-			expectedErrLength: 0,
-		},
-		{
-			conf: SyslogConf{
-				Protocol: "foo",
-				Port:     "514",
-			},
-			expectedErrLength: 1,
-		},
-		{
-			conf: SyslogConf{
-				Protocol: "invalid",
-				Port:     "-1",
-			},
-			expectedErrLength: 2,
-		},
-		{
-			conf: SyslogConf{
-				Protocol: "invalid",
-				Port:     "invalid",
-				Severity: "invalid",
-				Facility: "invalid",
-			},
-			expectedErrLength: 4,
-		},
-	}
-
-	for i, tt := range tests {
-		tt.conf.Enabled = true
-		errs := tt.conf.Validate()
-		if len(errs) != tt.expectedErrLength {
-			t.Errorf("test: %d, expected %d, actual %d", i, tt.expectedErrLength, len(errs))
-		}
-	}
-}
-
 func TestDistro_MajorVersion(t *testing.T) {
 	var tests = []struct {
 		in  Distro
