+++ config/syslog/syslogconf.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package config
+package syslog
 
 import (
 	"errors"
@@ -10,20 +10,8 @@ import (
 	"golang.org/x/xerrors"
 )
 
-// SyslogConf is syslog config
-type SyslogConf struct {
-	Protocol string `json:"-"`
-	Host     string `valid:"host" json:"-"`
-	Port     string `valid:"port" json:"-"`
-	Severity string `json:"-"`
-	Facility string `json:"-"`
-	Tag      string `json:"-"`
-	Verbose  bool   `json:"-"`
-	Enabled  bool   `toml:"-" json:"-"`
-}
-
 // Validate validates configuration
-func (c *SyslogConf) Validate() (errs []error) {
+func (c *Conf) Validate() (errs []error) {
 	if !c.Enabled {
 		return nil
 	}
@@ -52,7 +40,7 @@ func (c *SyslogConf) Validate() (errs []error) {
 }
 
 // GetSeverity gets severity
-func (c *SyslogConf) GetSeverity() (syslog.Priority, error) {
+func (c *Conf) GetSeverity() (syslog.Priority, error) {
 	if c.Severity == "" {
 		return syslog.LOG_INFO, nil
 	}
@@ -80,7 +68,7 @@ func (c *SyslogConf) GetSeverity() (syslog.Priority, error) {
 }
 
 // GetFacility gets facility
-func (c *SyslogConf) GetFacility() (syslog.Priority, error) {
+func (c *Conf) GetFacility() (syslog.Priority, error) {
 	if c.Facility == "" {
 		return syslog.LOG_AUTH, nil
 	}
